started app skeleton
